Start to implement Mazewar Shared State interface
This commits starts work on implementing the shared state object that
will be used to keep track of the current state of the game. The
external interface of the object is contained in the mwexternal.h file.
In addition, a few other files and objects were created for internal use
by the shared state object.