Add update method to the Mazewar State object
Added the mws_update method to the Mazewar State object. This method is
to be called on each iteration of the main game loop, and is used to
perform various operation for updating the game's state. Currently all
it does is update it's missiles positions. As a result of the new
mw_state method, the mw_missile object also received an mwm_update
method addition.