Add mws_render_wipe method to the Mazewar State
The way the rendering code is setup, objects need to clear their own
cells as they move around in the map. Thus, simply having a rendering
function is not enough, they also need a function to clear the cell they
were previously occupying. Thus the "render" event has been split up
into a "render_draw" event and a "render_wipe" event.

An example usage of this was implemented by the mw_missile class. It's
mwm_render_wipe method clears the cell it was previously occupying, and
it's mwm_render_draw method draws it's bitmap into the cell it is
currently occupying. It is perfectly legal to wipe the same cell which
will be drawn to, the draw command will overwrite the wipe commands as
draw occurs after wipe.