Missiles are now destroyed as they hit the walls
Missiles are now destroyed as they run into the maze walls. In order to
accomplish this, the mw_state object needed to inherit a new member
variable to use as the representation of the maze. Thus, a mw_state
object can now use this variable to determine where walls are placed
throughout the maze.