Fixed incorrect position passed to mwm_cons
The position passed to mwm_cons should actually be one cell in front of
the rat which fired it. This commit does just this. It also checks to
make sure a wall isn't directly in front of the rat. If this is the
case, there is no need to create a missile.