Add phases and elapsed time to mw_state_t
Added the mw_phase_t structure, and also added one as a member variable
to the mw_state_t class. As defined in the Mazewar Protocol
Specification, a client is to remain in the Discovery phase for a
certain amount of time prior to moving into the Active phase.

In addition to the mw_phase_t structure, logic was added to the
mw_state_t class to keep track of the elapsed time. This is currently
only used to transition from the Discovery phase, but could be used to
keep track of the timestamp required by the Protocol Spec.