Add state_pkt_timeout field to the mw_rat_t class
The Mazewar spec states that a state packet must be sent for each local
rat at least once every 500ms. Thus, a timeout timer was added to the
mw_rat_t class to keep track of this. Additional logic needs to be added
to check for when it triggers and perform some action (i.e. send a state
packet).