Add flag to control whether a rat sends packets
Only rats local to each mazewar client should broadcast packets to it's
peers. Thus, a flag was added to the mw_rat_t class to disable sending
packets by default. If a given rat is local to the client, this flag can
be set, and the rat will begin broadcasting it's state updates,
nickname, etc.