Add support for sending leaving packets
This commit adds support for sending leaving packets when a local
clients quit the game. Also, the leaving packet definition was slightly
modified to bump its size to a total of 64-bytes.

A bug was also fixed in the code sending nickname packets, the size
passed to sendto was incorrect.

Issue #25
Closes #19