Fix a bug in the posdir_[un]pack function
According to the Mazewar Protocol Spec, the position and direction need
to be packed into a 4-byte word, such that the low order bits hold the x
position, and the high order bits hold the direction.

Prior to this patch, the posdir_[un]pack function had this backwards. It
was storing x in the high bits, and direction in the low bits. This is
now fixed.