Remove mwr_list field from mw_rat_t
This commit removes the mwr_list field from the mw_rat_t object.
Instead, a new mw_rat_list_ent_t structure was created to hold this list
field. The reason for this is to allow per rat metadata to also be
stored in this rat list entry struct, e.g. such as timeout information.